blobfs: merge crate blobfs into crate rafs
Merge crate blobfs into crate rafs, to reduce number of crates.
