--- blobfs/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "nydus-blobfs"
-version = "0.2.0"
-description = "Blob object file system for Nydus Image Service"
-authors = ["The Nydus Developers"]
-license = "Apache-2.0 OR BSD-3-Clause"
-homepage = "https://nydus.dev/"
-repository = "https://github.com/dragonflyoss/image-service"
-edition = "2018"
-
-[dependencies]
-fuse-backend-rs = "^0.10.3"
-libc = "0.2"
-log = "0.4.8"
-serde = { version = "1.0.110", features = ["serde_derive", "rc"] }
-serde_json = "1.0.53"
-vm-memory = { version = "0.9" }
-
-nydus-error = { version = "0.2", path = "../error" }
-nydus-api = { version = "0.2", path = "../api" }
-nydus-rafs = { version = "0.2", path = "../rafs" }
-nydus-storage = { version = "0.6", path = "../storage", features = [
-    "backend-localfs",
-] }
-
-[features]
-virtiofs = ["fuse-backend-rs/virtiofs", "nydus-rafs/virtio-fs"]
-
-[package.metadata.docs.rs]
-all-features = true
-targets = ["x86_64-unknown-linux-gnu", "x86_64-apple-darwin"]
