+++ rafs/src/lib.rs
@@ -48,6 +48,8 @@ use std::sync::Arc;
 
 use crate::metadata::{RafsInodeExt, RafsSuper};
 
+#[cfg(feature = "virtio-fs")]
+pub mod blobfs;
 #[cfg(feature = "builder")]
 pub mod builder;
 pub mod fs;
